refactor: 패키지 위치, 클래스명 변경
