+++ src/main/java/roomescape/application/dto/MemberRequest.java
@@ -1,0 +1,6 @@
+package roomescape.application.dto;
+
+import roomescape.domain.member.MemberName;
+
+public record MemberRequest(MemberName name) {
+}
