+++ src/main/java/roomescape/application/dto/MemberResponse.java
@@ -1,0 +1,5 @@
+package roomescape.application.dto;
+
+public record MemberResponse(String name) {
+
+}
