+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -6,9 +6,9 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Objects;
+import roomescape.domain.member.MemberName;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.time.ReservationTime;
-import roomescape.domain.user.UserName;
 
 public class Reservation {
     private static final int IN_ADVANCE_RESERVATION_DAYS = 1;
@@ -16,7 +16,7 @@ public class Reservation {
     private final Long id;
     @NotBlank(message = "예약자명은 필수입니다.")
     @Size(min = 2, max = 10, message = "이름 길이는 2글자 이상, 10글자 이하여야 합니다.")
-    private final UserName name;
+    private final MemberName name;
 
     @NotBlank(message = "예약 날짜는 필수입니다.")
     private final LocalDate date;
@@ -32,14 +32,14 @@ public class Reservation {
     }
 
     public Reservation(Long id, String name, LocalDate date, ReservationTime time, Theme theme) {
-        this(id, new UserName(name), date, time, theme);
+        this(id, new MemberName(name), date, time, theme);
     }
 
     public Reservation(Long id, Reservation reservation) {
         this(id, reservation.name, reservation.date, reservation.time, reservation.theme);
     }
 
-    private Reservation(Long id, UserName name, LocalDate date, ReservationTime time, Theme theme) {
+    private Reservation(Long id, MemberName name, LocalDate date, ReservationTime time, Theme theme) {
         Objects.requireNonNull(date, "예약 날짜는 필수입니다.");
         validateReservationInAdvance(date, time.getStartAt());
